+++ src/nl/NG/Jetfightergame/Scenarios/SimplexCave.java
@@ -29,7 +29,7 @@ public class SimplexCave implements Touchable {
     private static final Material GROUND = Material.ROUGH;
     private static final float MERGE_FACTOR = 0.25f;
     // difference between minimum and maximum of OpenSimplexNoise
-    private static final double SIM_VAR = 2 * 0.852;
+    private static final float SIM_VAR = 2 * 0.852f;
 
     private Shape topGrid;
     private Shape bottomGrid;
@@ -56,10 +56,10 @@ public class SimplexCave implements Touchable {
         for(int x = 0; x < rows; x++){
             for(int y = 0; y < rows; y++){
                 // comparable to x and y
-                double xCoord = ((getRandom(scatter) + x) - rows/2);
-                double yCoord = ((getRandom(scatter) + y) - rows/2);
+                float xCoord = ((getRandom(scatter) + x) - rows/2);
+                float yCoord = ((getRandom(scatter) + y) - rows/2);
 
-                final double height = ((noise.eval(xCoord/pointDensity, yCoord/pointDensity) / SIM_VAR) + 0.5);
+                final float height =  ((float) noise.eval(xCoord/pointDensity, yCoord/pointDensity) / SIM_VAR + 0.5f);
                 grid[x][y] = new PosVector(xCoord, yCoord, height);
             }
         }
@@ -68,8 +68,8 @@ public class SimplexCave implements Touchable {
     }
 
     /** @return random between -scatter and scatter */
-    private static double getRandom(float scatter) {
-        return (2 * Settings.random.nextDouble() - 1f) * scatter;
+    private static float getRandom(float scatter) {
+        return (2 * Settings.random.nextFloat() - 1f) * scatter;
     }
 
     @Override
